requirements.txt
# Core Django dependencies
Django>=5.2.4,<6.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

# API and HTTP requests
requests>=2.31.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# API Documentation
drf-spectacular>=0.27.0

# Production server
gunicorn>=22.0.0

# Static files handling
whitenoise>=6.6.0
